+++ src/Httpful/UriResolver.php
@@ -268,13 +268,13 @@ final class UriResolver
         // as the first segment of a relative-path reference, as it would be mistaken for a scheme name.
         /* @phpstan-ignore-next-line | FP? */
         if ($relativePath === '' || \strpos(\explode('/', $relativePath, 2)[0], ':') !== false) {
-            $relativePath = "./${relativePath}";
+            $relativePath = "./{$relativePath}";
         } elseif ($relativePath[0] === '/') {
             if ($base->getAuthority() !== '' && $base->getPath() === '') {
                 // In this case an extra slash is added by resolve() automatically. So we must not add one here.
-                $relativePath = ".${relativePath}";
+                $relativePath = ".{$relativePath}";
             } else {
-                $relativePath = "./${relativePath}";
+                $relativePath = "./{$relativePath}";
             }
         }