+++ src/Httpful/ServerRequest.php
@@ -57,18 +57,18 @@ class ServerRequest extends Request implements ServerRequestInterface
     }
 
     /**
-     * @param string $attribute
+     * @param string $name
      * @param mixed  $default
      *
      * @return mixed|null
      */
-    public function getAttribute($attribute, $default = null)
+    public function getAttribute($name, $default = null)
     {
-        if (\array_key_exists($attribute, $this->attributes) === false) {
+        if (\array_key_exists($name, $this->attributes) === false) {
             return $default;
         }
 
-        return $this->attributes[$attribute];
+        return $this->attributes[$name];
     }
 
     /**
@@ -120,15 +120,15 @@ class ServerRequest extends Request implements ServerRequestInterface
     }
 
     /**
-     * @param string $attribute
+     * @param string $name
      * @param mixed  $value
      *
      * @return static
      */
-    public function withAttribute($attribute, $value): self
+    public function withAttribute($name, $value): self
     {
         $new = clone $this;
-        $new->attributes[$attribute] = $value;
+        $new->attributes[$name] = $value;
 
         return $new;
     }
@@ -138,7 +138,7 @@ class ServerRequest extends Request implements ServerRequestInterface
      *
      * @return ServerRequest|ServerRequestInterface
      */
-    public function withCookieParams(array $cookies)
+    public function withCookieParams(array $cookies): ServerRequestInterface
     {
         $new = clone $this;
         $new->cookieParams = $cookies;
@@ -151,7 +151,7 @@ class ServerRequest extends Request implements ServerRequestInterface
      *
      * @return ServerRequest|ServerRequestInterface
      */
-    public function withParsedBody($data)
+    public function withParsedBody($data): ServerRequestInterface
     {
         if (
             !\is_array($data)
@@ -174,7 +174,7 @@ class ServerRequest extends Request implements ServerRequestInterface
      *
      * @return ServerRequestInterface|static
      */
-    public function withQueryParams(array $query)
+    public function withQueryParams(array $query): ServerRequestInterface
     {
         $new = clone $this;
         $new->queryParams = $query;
@@ -187,7 +187,7 @@ class ServerRequest extends Request implements ServerRequestInterface
      *
      * @return ServerRequestInterface|static
      */
-    public function withUploadedFiles(array $uploadedFiles)
+    public function withUploadedFiles(array $uploadedFiles): ServerRequestInterface
     {
         $new = clone $this;
         $new->uploadedFiles = $uploadedFiles;
@@ -196,18 +196,18 @@ class ServerRequest extends Request implements ServerRequestInterface
     }
 
     /**
-     * @param string $attribute
+     * @param string $name
      *
      * @return static
      */
-    public function withoutAttribute($attribute): self
+    public function withoutAttribute($name): self
     {
-        if (\array_key_exists($attribute, $this->attributes) === false) {
+        if (\array_key_exists($name, $this->attributes) === false) {
             return $this;
         }
 
         $new = clone $this;
-        unset($new->attributes[$attribute]);
+        unset($new->attributes[$name]);
 
         return $new;
     }