+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 3.0.0 (2023-07-20)
+
+- allow to use "psr/http-message" 2.0.* 
+- allow to use "psr/log" 2.0.* || 3.0.*
+
+breaking change:
+- fixed "Response->hasBody()", now if will return `false` for an empty body
+- "Stream->getContents()" now returns always a string, if we need the old behaviors, use can use "Stream->getContentsUnserialized()"
+- "psr/http-message" v2 has return types, so you need to use them too, if you extend one of this classes
+
 ## 2.4.9 (2023-07-15)
 
 - use "ReturnTypeWillChange" to ignore return type changes from PHP >= 8.1